stages/sink/requirements.txt
kafka-python==2.0.*